backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foodflow-os-backend"
version = "0.1.0"
description = "FoodFlow OS Backend - AI-powered operating system for food manufacturing"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "FoodFlow OS Team", email = "team@foodflow.com"}
]
keywords = ["food", "manufacturing", "ai", "fsq", "traceability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    
    # Redis
    "redis[hiredis]>=5.0.1",
    "aioredis>=2.0.1",
    
    # HTTP Client
    "httpx>=0.25.2",
    "aiohttp>=3.9.1",
    
    # Data Validation & Serialization
    "email-validator>=2.1.0",
    "python-dateutil>=2.8.2",
    
    # Background Tasks
    "celery[redis]>=5.3.4",
    
    # Logging & Monitoring
    "structlog>=23.2.0",
    "python-json-logger>=2.0.7",
    
    # OpenTelemetry
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.42b0",
    
    # ML/AI
    "scikit-learn>=1.3.2",
    "numpy>=1.26.2",
    "pandas>=2.1.4",
    "xgboost>=2.0.3",
    "lightgbm>=4.1.0",
    "prophet>=1.1.5",
    
    # LLM & RAG
    "openai>=1.6.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "tiktoken>=0.5.2",
    
    # Vector Search
    "pgvector>=0.2.4",
    
    # Time Series
    "statsmodels>=0.14.1",
    
    # Optimization
    "pulp>=2.7.0",
    "ortools>=9.8.3296",
    
    # Utilities
    "python-slugify>=8.0.1",
    "pytz>=2023.3",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "hypothesis>=6.92.1",
    "faker>=21.0.0",
    
    # Code Quality
    "ruff>=0.1.8",
    "black>=23.12.1",
    "mypy>=1.7.1",
    "isort>=5.13.2",
    
    # Type Stubs
    "types-python-dateutil>=2.8.19",
    "types-redis>=4.6.0",
    "types-passlib>=1.7.7",
    
    # Development Tools
    "ipython>=8.18.1",
    "ipdb>=0.13.13",
    "watchfiles>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/OPSC2025/foodflow-os"
Documentation = "https://github.com/OPSC2025/foodflow-os/tree/main/docs"
Repository = "https://github.com/OPSC2025/foodflow-os"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_untyped_calls = false

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101"]
